Implement graceful shutdown in runner

+++ packages/tests/src/shared/peertube-runner-process.ts
@@ -89,13 +89,27 @@ export class PeerTubeRunnerProcess {
     return stdout
   }
 
+  // ---------------------------------------------------------------------------
+
+  gracefulShutdown () {
+    const args = [ 'graceful-shutdown', ...this.buildIdArg() ]
+
+    return this.runCommand(this.getRunnerPath(), args)
+  }
+
+  hasCorrectlyExited () {
+    return this.app.exitCode === 0
+  }
+
   kill () {
-    if (!this.app) return
+    if (!this.app || this.app.exitCode !== null) return
 
     process.kill(this.app.pid)
 
     this.app = null
   }
+
+  // ---------------------------------------------------------------------------
 
   getId () {
     return 'test-' + this.server.internalServerNumber
